Removed CV training function
This was removed due to it taking in an already formed classifier.
This meant that the scikit SGDClassifier partial_fit function was
being called across CV folds. This caused later folds to benefit from
the partial fitting of previous folds, thereby breaking the intended
separation of folds for training and testing. It was easier to remove
this and rely on user implemented CV than to fix it.